added

+++ src/main/java/com/evry/service/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package com.evry.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,7 +11,7 @@ import com.evry.repository.UserRepository;
 
 @Service
 public class UserServiceImpl implements UserService {
-	
+
 	@Autowired
 	private UserRepository repository;
 
@@ -20,13 +21,13 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public User updateUser(User user) {
-		return repository.save(user);
-	}
-
-	@Override
-	public void deleteUserById(long id) {
-		 repository.deleteById(id);
+	public void deleteUserById(int id) {
+		Optional<User> newUser = repository.findById(id);
+		if(newUser.isPresent()) {
+			repository.deleteById(id);
+		}else {
+			throw new RuntimeException("User with id { "+ id + "} not found ");
+		}
 	}
 
 	@Override
@@ -36,7 +37,13 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public User getUserById(int id) {
-		return repository.findById((long) id).get();
+		Optional<User> newUser = repository.findById(id);
+
+		if (newUser.isPresent()) {
+			return newUser.get();
+		} else {
+			throw new RuntimeException("User with id {"+ id +"} not found ");
+		}
 	}
 
 	@Override
@@ -44,6 +51,17 @@ public class UserServiceImpl implements UserService {
 		return repository.findAll();
 	}
 
-	
+	@Override
+	public User updateUser(int id, User u) {
+		User user = repository.findById(id).get();
+		user.setFirstName(u.getFirstName());
+		user.setLastName(u.getLastName());
+		user.setEmail(u.getEmail());
+		user.setGender(u.getGender());
+		user.setPassword(u.getPassword());
+
+		user.setAddress(u.getAddress());
+		return repository.save(user);
+	}
 
 }
